Stratum moved to own directory, cosmetic fixes
